refactor: migrate CIMEError imports to CIME.core.exceptions
Replace the CIMEError class definition in utils.py with a re-export from
CIME.core.exceptions for backward compatibility. Update 19 internal
callers to import CIMEError directly from CIME.core.exceptions.

External consumers using 'from CIME.utils import CIMEError' or
'utils.CIMEError' continue to work unchanged via the re-export.
