+++ CIME/case/case_run.py
@@ -5,7 +5,8 @@ case_run is a member of Class Case
 from CIME.XML.standard_module_setup import *
 from CIME.config import Config
 from CIME.utils import gzip_existing_file, new_lid
-from CIME.utils import run_sub_or_cmd, safe_copy, model_log, CIMEError
+from CIME.core.exceptions import CIMEError
+from CIME.utils import run_sub_or_cmd, safe_copy, model_log
 from CIME.utils import batch_jobid, is_comp_standalone
 from CIME.status import append_status, run_and_log_case_status
 from CIME.get_timing import get_timing
