+++ CIME/tests/test_unit_xml_env_batch.py
@@ -6,7 +6,8 @@ import tempfile
 from contextlib import ExitStack
 from unittest import mock
 
-from CIME.utils import CIMEError, expect
+from CIME.core.exceptions import CIMEError
+from CIME.utils import expect
 from CIME.XML.env_batch import EnvBatch, get_job_deps
 from CIME.XML.env_workflow import EnvWorkflow
 from CIME.BuildTools.configure import FakeCase
@@ -513,8 +514,7 @@ class TestXMLEnvBatch(unittest.TestCase):
 
     def test_get_submit_args_job_queue(self):
         with tempfile.NamedTemporaryFile() as tfile:
-            tfile.write(
-                b"""<?xml version="1.0"?>
+            tfile.write(b"""<?xml version="1.0"?>
 <file id="env_batch.xml" version="2.0">
   <header>
       These variables may be changed anytime during a run, they
@@ -547,8 +547,7 @@ class TestXMLEnvBatch(unittest.TestCase):
     </queues>
   </batch_system>
 </file>
-"""
-            )
+""")
 
             tfile.seek(0)
 
@@ -570,8 +569,7 @@ class TestXMLEnvBatch(unittest.TestCase):
     @mock.patch.dict(os.environ, {"TEST": "GOOD"})
     def test_get_submit_args(self):
         with tempfile.NamedTemporaryFile() as tfile:
-            tfile.write(
-                b"""<?xml version="1.0"?>
+            tfile.write(b"""<?xml version="1.0"?>
 <file id="env_batch.xml" version="2.0">
   <header>
       These variables may be changed anytime during a run, they
@@ -628,8 +626,7 @@ class TestXMLEnvBatch(unittest.TestCase):
     </queues>
   </batch_system>
 </file>
-"""
-            )
+""")
 
             tfile.seek(0)
